<chore> create Reply ToogleButton for  reusability
