+++ front/src/components/comment/comment.tsx
@@ -1,11 +1,11 @@
 "use client"
-import { Flex,Box,Text,Spacer,IconButton, Button } from "@chakra-ui/react"
-import { CommentType , CommentClass } from "@/atoms/comment";
-import { SlPlus , SlMinus } from "react-icons/sl";
+import { Flex,Text,Spacer, Button } from "@chakra-ui/react"
+import {  CommentClass } from "@/atoms/comment";
 import React, { useState } from "react";
 import { customColors } from "@/utils/chakra/customColors";
 import { mdFontSize } from "@/utils/chakra/fonts";
 import CommentMenu from "./commentMenu";
+import ReplyToggleButton from "./replyToggleButton";
 
 
 type CommentProps = {
@@ -25,12 +25,15 @@ const Comment : React.FC<CommentProps> = ({
     isShowReply,
     handleShowReplyIconClicked
 }) => {
+    
     const comment = commentInstance.getComment();
     const content = comment.content;
     const author = comment.author.nickname || comment.author.email;
     const likeCount = comment.likeCount;
     const timeAgo = comment.createdAt.toString();
     const buttonColor = customColors.black[200];
+    const hasReplies = commentInstance.isHaveReply();
+
 
     return (
         <Flex 
@@ -40,23 +43,20 @@ const Comment : React.FC<CommentProps> = ({
             px={2}
             py={1}
         >
-            <Box width={"10%"}>
+            <Flex width={"10%"} justifyContent={"space-between"} flexDir={"column"} >
                 <Text>{author}</Text>
                 <Spacer/>
-                <IconButton
-                    hidden= { !commentInstance.isHaveReply() }
-                    isRound={true}
-                    aria-label="Reply Comment"
-                    icon={isShowReply ? <SlMinus/> : <SlPlus/>}
-                    onClick={handleShowReplyIconClicked}
-                />
-            </Box>
+                {hasReplies&&<ReplyToggleButton
+                    isShowReply={isShowReply}
+                    handleShowReplyIconClicked={handleShowReplyIconClicked}
+                    fontSize="1.5rem"
+                />}
+            </Flex>
             <Flex flexDir={"column"} gap = "0.5rem">
                 <Flex>
                     <Text>{timeAgo}</Text>
                     <Spacer />
                     <CommentMenu commentInstance={commentInstance}/>
-
                 </Flex>
                 <Text fontSize={mdFontSize}>{content}</Text>
                 <Flex gap="0.5rem">
